Conflicts:
	flexo-foundation/src/main/java/org/openflexo/foundation/DefaultFlexoServiceManager.java
	flexo-foundation/src/main/java/org/openflexo/foundation/fml/rm/ViewPointResourceImpl.java
	flexo-foundation/src/main/java/org/openflexo/foundation/resource/DefaultResourceCenterService.java
	flexo-foundation/src/main/java/org/openflexo/foundation/resource/JarResourceCenter.java
	flexo-foundation/src/main/java/org/openflexo/foundation/resource/RemoteResourceCenter.java
	flexo-foundation/src/test/java/org/openflexo/foundation/OpenFlexoTestCaseWithGit.java
	flexo-foundation/src/test/java/org/openflexo/foundation/OpenflexoProjectAtRunTimeTestCase.java
	flexo-foundation/src/test/java/org/openflexo/foundation/OpenflexoTestCase.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestControlGraph.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestCreateFlexoConcept.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestCreateViewPoint.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestCreateViewPointInProject.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestCreateVirtualModel.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestDeleteAction.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestDeleteFlexoConceptObjects.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestFMLBindingModelManagement.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestFMLBindingModelManagement2.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestFlexoConceptInheritance.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestFlexoConceptInstanceType.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/TestFlexoRoleCardinality.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/rt/TestCreateView.java
	flexo-foundation/src/test/java/org/openflexo/foundation/fml/rt/TestCreateVirtualModelInstance.java
	flexo-foundation/src/test/java/org/openflexo/foundation/rm/TestDefaultRCServiceLoadRCInClassPath.java
	flexo-foundation/src/test/java/org/openflexo/foundation/task/TestFlexoTaskManager.java
	flexo-ui/src/test/java/org/openflexo/OpenflexoProjectAtRunTimeTestCaseWithGUI.java
	flexo-ui/src/test/java/org/openflexo/OpenflexoTestCaseWithGUI.java
	flexo-ui/src/test/java/org/openflexo/TestPreferencesService.java
	flexo-ui/src/test/java/org/openflexo/fib/TestPreferencesDialog.java

+++ flexo-ui-test/src/main/java/org/openflexo/gina/test/OpenflexoTestCaseWithGUI.java
@@ -83,7 +83,7 @@ public abstract class OpenflexoTestCaseWithGUI extends OpenflexoTestCase {
 			FlexoResourceCenterService RCService = serviceManager.getResourceCenterService();
 			List<FlexoResourceCenter> listRC = RCService.getResourceCenters();
 			for (FlexoResourceCenter rc : listRC) {
-				if (rc instanceof DirectoryResourceCenter) {
+				if (rc instanceof DirectoryResourceCenter && !rc.getResourceCenterEntry().isSystemEntry()) {
 					File RCDirectory = ((DirectoryResourceCenter) rc).getDirectory();
 					RCDirectory.deleteOnExit();
 				}
@@ -117,7 +117,13 @@ public abstract class OpenflexoTestCaseWithGUI extends OpenflexoTestCase {
 
 	protected static ApplicationContext instanciateTestServiceManager(final boolean generateCompoundTestResourceCenter) {
 		serviceManager = new TestApplicationContext(generateCompoundTestResourceCenter);
-		resourceCenter = (DirectoryResourceCenter) serviceManager.getResourceCenterService().getResourceCenters().get(0);
+		for (FlexoResourceCenter rc : serviceManager.getResourceCenterService().getResourceCenters()) {
+			// Select the first directory ResourceCenter
+			if (rc instanceof DirectoryResourceCenter && !rc.getResourceCenterEntry().isSystemEntry()) {
+				resourceCenter = (DirectoryResourceCenter) rc;
+				break;
+			}
+		}
 		return serviceManager;
 	}
 
@@ -128,7 +134,13 @@ public abstract class OpenflexoTestCaseWithGUI extends OpenflexoTestCase {
 			serviceManager
 					.activateTechnologyAdapter(serviceManager.getTechnologyAdapterService().getTechnologyAdapter(technologyAdapterClass));
 		}
-		resourceCenter = (DirectoryResourceCenter) serviceManager.getResourceCenterService().getResourceCenters().get(0);
+		for (FlexoResourceCenter rc : serviceManager.getResourceCenterService().getResourceCenters()) {
+			// Select the first directory ResourceCenter
+			if (rc instanceof DirectoryResourceCenter) {
+				resourceCenter = (DirectoryResourceCenter) rc;
+				break;
+			}
+		}
 		return serviceManager;
 	}
